Upgrade depencies (rspec-maven-plugin, rspec) + specs refactor
Travis is currently using maven 3.2.5 and the rspec-maven-plugin version we
were using doesn't work with that. So it was upgraded to the latest version.

The version of the rspec gem we were using, apparently, doesn't work with the
new version of the rspec-maven-plugin. So it was upgraded as well.

Specs were refactored to comply with the new rspec api.